fix decompose rule

+++ python/torch_mlir_e2e_test/test_suite/basic.py
@@ -3738,7 +3738,7 @@ class ScalarTensorModule(torch.nn.Module):
         None,
     ])
     def forward(self):
-        scalar = torch.ops.aten.scalar_tensor(1.0, dtype=torch.float64)
+        scalar = torch.ops.aten.scalar_tensor(1.0, dtype=torch.float32)
         return scalar
